Track inheritance column to avoid STI type loss
When using STI, auditable_type always corresponds to the base model.
Besides that, the inheritance column is not tracked. Therefore, after an
inherited model was destroyed, it is impossible to infer it's type from
the audit.

This removes the inheritance column from the default ignored attributes
in order to keep track of the type, which will instantiate the proper
class when using the #revision call.