Use saved_changes instead of changed_attributes
This commit follows the guidance of a deprecation warning that is
generated under Rails 5.1:

DEPRECATION WARNING: The behavior of `changed_attributes` inside of
after callbacks will be changing in the next version of Rails. The new
return value will reflect the behavior of calling the method after
`save` returned (e.g. the opposite of what it returns now). To maintain
the current behavior, use `saved_changes.transform_values(&:first)` instead.